README + re order
