+++ Back-END/Delivery/controllers/logger.js
@@ -1,14 +1,17 @@
-const { mongoose } = require('./dbMongo');
+const Logs = require("../models/logger");
 
-var sc = new mongoose.Schema({
-  timestamp: String,
-  service: String,
-  action_type: String,
-  route: String,
-  success: Boolean,
-  message: String,
-});
-
-const Logs = mongoose.model('logs', sc);
-
-module.exports = Logs;
+module.exports = {
+    logaction: async function  (action_type, route, success = false, message = "") {
+        const current = new Date();
+        var options = { day: '2-digit', month: '2-digit', year: 'numeric', hour: '2-digit', minute: '2-digit', second: '2-digit' };
+        const line = new Logs({
+            timestamp: current.toLocaleDateString('fr', options),
+            service: "Delivery",
+            action_type: action_type,
+            route: route,
+            success: success,
+            message: message,
+        })
+        await line.save();
+    }
+}
