--- Back-END/Delivery/middlewares/logger.js
@@ -1,17 +1,0 @@
-const Logs = require("../controllers/logger");
-
-module.exports = {
-    logaction: async function  (action_type, route, success = false, message = "") {
-        const current = new Date();
-        var options = { day: '2-digit', month: '2-digit', year: 'numeric', hour: '2-digit', minute: '2-digit', second: '2-digit' };
-        const line = new Logs({
-            timestamp: current.toLocaleDateString('fr', options),
-            service: "Delivery",
-            action_type: action_type,
-            route: route,
-            success: success,
-            message: message,
-        })
-        await line.save();
-    }
-}
